# soroban-rpc: Fix caching of GetLatestLedgerSequence